requirements.txt
mysql-connector-python
django==2.1